+++ lxd/devices.go
@@ -1,9 +1,11 @@
 package main
 
 import (
+	"database/sql"
 	"fmt"
 
 	"github.com/lxc/lxd/shared"
+	_ "github.com/mattn/go-sqlite3"
 )
 
 func DeviceToLxc(d shared.Device) ([][]string, error) {
@@ -67,4 +69,140 @@ func DeviceToLxc(d shared.Device) ([][]string, error) {
 	default:
 		return nil, fmt.Errorf("Bad device type")
 	}
+}
+
+func ValidDeviceType(t string) bool {
+	switch t {
+	case "unix-char":
+		return true
+	case "unix-block":
+		return true
+	case "nic":
+		return true
+	case "disk":
+		return true
+	case "none":
+		return true
+	default:
+		return false
+	}
+}
+
+func ValidDeviceConfig(t, k, v string) bool {
+	if k == "type" {
+		return false
+	}
+	switch t {
+	case "unix-char":
+		switch k {
+		case "path":
+			return true
+		case "major":
+			return true
+		case "minor":
+			return true
+		case "uid":
+			return true
+		case "gid":
+			return true
+		case "mode":
+			return true
+		default:
+			return false
+		}
+	case "unix-block":
+		switch k {
+		case "path":
+			return true
+		case "major":
+			return true
+		case "minor":
+			return true
+		case "uid":
+			return true
+		case "gid":
+			return true
+		case "mode":
+			return true
+		default:
+			return false
+		}
+	case "nic":
+		switch k {
+		case "parent":
+			return true
+		case "name":
+			return true
+		case "hwaddr":
+			return true
+		case "mtu":
+			return true
+		case "nictype":
+			if v != "bridged" && v != "" {
+				return false
+			}
+			return true
+		default:
+			return false
+		}
+	case "disk":
+		switch k {
+		case "path":
+			return true
+		case "source":
+			return true
+		case "readonly", "optional":
+			return true
+		default:
+			return false
+		}
+	case "none":
+		return false
+	default:
+		return false
+	}
+}
+
+func AddDevices(tx *sql.Tx, w string, cId int, devices shared.Devices) error {
+	str1 := fmt.Sprintf("INSERT INTO %ss_devices (%s_id, name, type) VALUES (?, ?, ?)", w, w)
+	stmt1, err := tx.Prepare(str1)
+	if err != nil {
+		return err
+	}
+	defer stmt1.Close()
+	str2 := fmt.Sprintf("INSERT INTO %ss_devices_config (%s_device_id, key, value) VALUES (?, ?, ?)", w, w)
+	stmt2, err := tx.Prepare(str2)
+	if err != nil {
+		return err
+	}
+	defer stmt2.Close()
+	for k, v := range devices {
+		if !ValidDeviceType(v["type"]) {
+			return fmt.Errorf("Invalid device type %s\n", v["type"])
+		}
+		result, err := stmt1.Exec(cId, k, v["type"])
+		if err != nil {
+			return err
+		}
+		id64, err := result.LastInsertId()
+		if err != nil {
+			return fmt.Errorf("Error inserting device %s into database", k)
+		}
+		// TODO: is this really int64? we should fix it everywhere if so
+		id := int(id64)
+		for ck, cv := range v {
+			if ck == "type" {
+				continue
+			}
+			if !ValidDeviceConfig(v["type"], ck, cv) {
+				return fmt.Errorf("Invalid device config %s %s\n", ck, cv)
+			}
+			_, err = stmt2.Exec(id, ck, cv)
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
 }
