+++ shared/container.go
@@ -1,13 +1,43 @@
 package shared
 
-import (
-	"database/sql"
-	"fmt"
-	"net"
+/*
+ * N.B. State is copied from lxc.State, but we re-export it here so that
+ * client libraries don't have to import go-lxc and thus link against liblxc
+ * for just some constants.
+ */
 
-	_ "github.com/mattn/go-sqlite3"
-	"gopkg.in/lxc/go-lxc.v2"
+// State type specifies possible container states.
+type State int
+
+const (
+	// STOPPED means container is not running
+	STOPPED State = iota + 1
+	// STARTING means container is starting
+	STARTING
+	// RUNNING means container is running
+	RUNNING
+	// STOPPING means container is stopping
+	STOPPING
+	// ABORTING means container is aborting
+	ABORTING
+	// FREEZING means container is freezing
+	FREEZING
+	// FROZEN means containe is frozen
+	FROZEN
+	// THAWED means container is thawed
+	THAWED
 )
+
+var StateMap = map[string]State{
+	"STOPPED":  STOPPED,
+	"STARTING": STARTING,
+	"RUNNING":  RUNNING,
+	"STOPPING": STOPPING,
+	"ABORTING": ABORTING,
+	"FREEZING": FREEZING,
+	"FROZEN":   FROZEN,
+	"THAWED":   THAWED,
+}
 
 type Ip struct {
 	Interface string `json:"interface"`
@@ -16,43 +46,10 @@ type Ip struct {
 }
 
 type ContainerStatus struct {
-	State     string    `json:"status"`
-	StateCode lxc.State `json:"status_code"`
-	Init      int       `json:"init"`
-	Ips       []Ip      `json:"ips"`
-}
-
-func getIps(c *lxc.Container) []Ip {
-	ips := []Ip{}
-	names, err := c.Interfaces()
-	if err != nil {
-		return ips
-	}
-	for _, n := range names {
-		addresses, err := c.IPAddress(n)
-		if err != nil {
-			continue
-		}
-		for _, a := range addresses {
-			ip := Ip{Interface: n, Address: a}
-			if net.ParseIP(a).To4() == nil {
-				ip.Protocol = "IPV6"
-			} else {
-				ip.Protocol = "IPV4"
-			}
-			ips = append(ips, ip)
-		}
-	}
-	return ips
-}
-
-func NewStatus(c *lxc.Container, state lxc.State) ContainerStatus {
-	status := ContainerStatus{State: state.String(), StateCode: state}
-	if state == lxc.RUNNING {
-		status.Init = c.InitPid()
-		status.Ips = getIps(c)
-	}
-	return status
+	State     string `json:"status"`
+	StateCode State  `json:"status_code"`
+	Init      int    `json:"init"`
+	Ips       []Ip   `json:"ips"`
 }
 
 type Device map[string]string
@@ -92,8 +89,8 @@ func (c *ContainerState) BriefState() BriefContainerState {
 	return retstate
 }
 
-func (c *ContainerState) State() lxc.State {
-	return lxc.StateMap[c.Status.State]
+func (c *ContainerState) State() State {
+	return StateMap[c.Status.State]
 }
 
 type ContainerInfo struct {
@@ -115,140 +112,4 @@ type ProfileConfig struct {
 	Name    string            `json:"name"`
 	Config  map[string]string `json:"config"`
 	Devices Devices           `json:"devices"`
-}
-
-func ValidDeviceType(t string) bool {
-	switch t {
-	case "unix-char":
-		return true
-	case "unix-block":
-		return true
-	case "nic":
-		return true
-	case "disk":
-		return true
-	case "none":
-		return true
-	default:
-		return false
-	}
-}
-
-func ValidDeviceConfig(t, k, v string) bool {
-	if k == "type" {
-		return false
-	}
-	switch t {
-	case "unix-char":
-		switch k {
-		case "path":
-			return true
-		case "major":
-			return true
-		case "minor":
-			return true
-		case "uid":
-			return true
-		case "gid":
-			return true
-		case "mode":
-			return true
-		default:
-			return false
-		}
-	case "unix-block":
-		switch k {
-		case "path":
-			return true
-		case "major":
-			return true
-		case "minor":
-			return true
-		case "uid":
-			return true
-		case "gid":
-			return true
-		case "mode":
-			return true
-		default:
-			return false
-		}
-	case "nic":
-		switch k {
-		case "parent":
-			return true
-		case "name":
-			return true
-		case "hwaddr":
-			return true
-		case "mtu":
-			return true
-		case "nictype":
-			if v != "bridged" && v != "" {
-				return false
-			}
-			return true
-		default:
-			return false
-		}
-	case "disk":
-		switch k {
-		case "path":
-			return true
-		case "source":
-			return true
-		case "readonly", "optional":
-			return true
-		default:
-			return false
-		}
-	case "none":
-		return false
-	default:
-		return false
-	}
-}
-
-func AddDevices(tx *sql.Tx, w string, cId int, devices Devices) error {
-	str1 := fmt.Sprintf("INSERT INTO %ss_devices (%s_id, name, type) VALUES (?, ?, ?)", w, w)
-	stmt1, err := tx.Prepare(str1)
-	if err != nil {
-		return err
-	}
-	defer stmt1.Close()
-	str2 := fmt.Sprintf("INSERT INTO %ss_devices_config (%s_device_id, key, value) VALUES (?, ?, ?)", w, w)
-	stmt2, err := tx.Prepare(str2)
-	if err != nil {
-		return err
-	}
-	defer stmt2.Close()
-	for k, v := range devices {
-		if !ValidDeviceType(v["type"]) {
-			return fmt.Errorf("Invalid device type %s\n", v["type"])
-		}
-		result, err := stmt1.Exec(cId, k, v["type"])
-		if err != nil {
-			return err
-		}
-		id64, err := result.LastInsertId()
-		if err != nil {
-			return fmt.Errorf("Error inserting device %s into database", k)
-		}
-		// TODO: is this really int64? we should fix it everywhere if so
-		id := int(id64)
-		for ck, cv := range v {
-			if ck == "type" {
-				continue
-			}
-			if !ValidDeviceConfig(v["type"], ck, cv) {
-				return fmt.Errorf("Invalid device config %s %s\n", ck, cv)
-			}
-			_, err = stmt2.Exec(id, ck, cv)
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
 }
